+++ src/organisations/admin/organisation-versions.controller.ts
@@ -4,6 +4,7 @@ import {
   Get,
   Render,
   Param,
+  Put,
   Post,
   Res,
   Req,
@@ -21,6 +22,7 @@ import { OrganisationVersionsService } from '../organisation-versions.service';
 import { RequestWithAppSession } from '../../common/interfaces/request-with-app-session.interface';
 
 import { OrganisationVersion } from '../organisation-version.entity';
+import { Organisation } from '../organisation.entity';
 
 import { ShowTemplate } from '../interfaces/show-template.interface';
 
@@ -77,11 +79,16 @@ export class OrganisationVersionsController {
     @Param('organisationId') organisationId: string,
     @Param('versionId') versionId: string,
   ): Promise<ShowTemplate> {
-    const organisation =
+    const version =
       await this.organisationVersionsService.findByIdWithOrganisation(
         organisationId,
         versionId,
       );
+
+    const organisation = Organisation.withVersion(
+      version.organisation,
+      version,
+    );
 
     const organisationSummaryPresenter = new OrganisationSummaryPresenter(
       organisation,
@@ -89,5 +96,22 @@ export class OrganisationVersionsController {
     );
 
     return organisationSummaryPresenter.present();
+  }
+
+  @Put(':organisationId/versions/:versionId/publish')
+  @Render('admin/organisations/versions/publish')
+  async publish(
+    @Param('organisationId') organisationId: string,
+    @Param('versionId') versionId: string,
+  ): Promise<Organisation> {
+    const version =
+      await this.organisationVersionsService.findByIdWithOrganisation(
+        organisationId,
+        versionId,
+      );
+
+    await this.organisationVersionsService.publish(version);
+
+    return version.organisation;
   }
 }